requirements.txt
manim
python-pptx